services/blockchain_service/pyproject.toml
[tool.poetry]
name = "blockchain-service"
version = "0.1.0"
description = ""
authors = ["Михаил <58465577+bubled33@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
libs = {path = "../../libs", develop = true}
pydantic-settings = "^2.12.0"
uvicorn = "^0.38.0"
web3 = "^7.14.0"
redis = "^7.1.0"
yoyo-migrations = "^9.0.0"
asyncpg = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
